requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
async-lru==2.0.4
attrs==23.2.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dict2xml==1.7.5
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
f5-icontrol-rest==1.3.13
frozenlist==1.4.1
future==1.0.0
genie==24.3
genie.libs.clean==24.3
genie.libs.conf==24.3
genie.libs.filetransferutils==24.3
genie.libs.health==24.3
genie.libs.ops==24.3
genie.libs.parser==24.3
genie.libs.sdk==24.3
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
idna==3.6
Jinja2==3.1.3
jsonpickle==3.0.3
junit-xml==1.9
lxml==4.9.4
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
ncclient==0.6.15
netaddr==0.10.1
netmiko==4.3.0
ntc_templates==4.4.0
openpyxl==3.1.2
packaging==24.0
paramiko==3.4.0
pathspec==0.12.1
platformdirs==4.2.0
ply==3.11
prettytable==3.10.0
protobuf==5.26.1
psutil==5.9.8
pyasn1==0.4.8
pyats==24.3
pyats.aereport==24.3
pyats.aetest==24.3
pyats.async==24.3
pyats.connections==24.3
pyats.datastructures==24.3
pyats.easypy==24.3
pyats.kleenex==24.3
pyats.log==24.3
pyats.reporter==24.3
pyats.results==24.3
pyats.tcl==24.3
pyats.topology==24.3
pyats.utils==24.3
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
pyftpdlib==1.5.9
pylama==8.4.1
PyNaCl==1.5.0
pyserial==3.5
pysmi-lextudio==1.4.3
pysnmp-lextudio==5.0.29
pysnmpcrypto==0.0.4
python-dotenv==1.0.1
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0.1
requests==2.31.0
rest.connector==24.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.14.5
setuptools==69.2.0
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
textfsm==1.1.3
tftpy==0.8.0
tqdm==4.66.2
unicon==24.3
unicon.plugins==24.3
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
xmltodict==0.13.0
yamllint==1.35.1
yang.connector==24.3
yarl==1.9.4
